phase2_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
python-multipart
google-generativeai
typing_extensions
pypdf
python-docx
scikit-learn
numpy
supabase